Rename: commos 폴더 추가로 파일 경로 수정
