+++ fe/.storybook/preview.ts
@@ -1,6 +1,7 @@
 import { withThemeFromJSXProvider } from '@storybook/addon-styling';
 import { ThemeProvider } from 'styled-components';
 import { GlobalStyle, theme } from '../src/styles/GlobalStyle';
+import '@styles/index.css';
 
 export const decorators = [
   withThemeFromJSXProvider({
