+++ fe/src/components/commons/Icon/Icon.stories.ts
